feat: add splitting to components
